Update BodyChanging.java
Added the Hair Type change to the Slime Transformations, allowing a character to be changed to a 'Greater' whatever. This permits greater Angels without cheating, as well as Greater Morphs via the Biojuice Canister and a Slime Elixer (Arcane Boosted to Flesh). Compiled and tested in-game (2.6.5); successful for Angel, Equine, Alligator and Wolf Morphs.