Add configuration of NPCs sexual orientation encounter rates
This change adds an option for players to set the encounter rates for
the sexual orientation of random NPCs, in the same way as gender
encounter rates.

The generation function uses both the generic player-defined rates and
call-specific rates when computing the final orientation. The specific
rates can be used for race/femininity-dependent rates. For example,
harpies were previously hardcoded as being almost always gynephilic, and
now use the generation function with unbalanced specific rates.