More detailed error messages during mod loading
There are a number of errors that can occur when importing a mod file
which give little idea as to what actually went wrong. This change
adds the filename to importing AbstractClothingType, AbstractTattooType,
and AbstractWeaponType, and adds the exception message to each error
case. Hopefully this will help mod authors figure out what exactly the
problem is.