Increased variety for Name.java
Increased the name lists by a considerable degree, increased the variety (fewer Carls!), and made the "prostitute" name generator occasionally produce "regular human" names (50% of the time) and vice versa (10% of the time). That last change can be undone if it's an issue, like if it'd be better to handle that elsewhere in the code.