Fix for over-eager comment removal
The game's text parser recently got the ability to parse out comments,
denoted by //, but it was a bit overzealous. It would consume things that
shouldn't be comments, such as urls like http://lilithsthrone.blogspot.co.uk,
leaving just http: hanging, and obliterating everything else on that line.
This caused the start screen to look weird, and nearly every svg to be rendered
at the wrong size.

I've put in a fix for this by adding a negative lookbehind for ':'. Other
options could be to require comments are on their own line (not ideal),
comments begin with ' //' (that is a space, then two slashes), or similar,
but I think this is the least restrictive option.