POC implementation of moving getTextFilePath to WorldType or WorldRegion
Right now this might be overkill because we only have offspring-encounter-XMLs for two locations: `harpyNests.xml` and `dominionAlleyway.xml`. With this PR needing one for the bat cave and the submission tunnels needing another one there are four. Considering that we need at least one for each region of the world map this would result in a loong if-elseif-else-chain. To avoid that this relocates specifying the `textFilePath` per WorldType and/or per WorldRegion.