Update Penis.java
Penis girth is really big on absurdly long penises, and changing the linear function to an exponential one would change that. The net effect is that really short penises would be wider by default, and really long ones would be thinner. The changes are greatest for the really long penises. The end result of this change is that from 1 to 9cm, the base girth is slightly larger than the original linear function, and 10cm is the inflection point where the girth (diameter) increase slows down. At 1cm girth becomes 0.5cm, at 5cm it's 1.5cm, at 10cm it's 2.5cm, 15cm : 3.3cm, 30cm : 5.4, 60cm : 8.8cm, and at 100 cm : 12.6cm. Fat and Thin would be double and half these numbers like usual.

Changed the math for diameter from length * 0.25 to length in mm ^ 0.7

No new graphical assets are required

I've not tested this change in-game