Fix sorting comparator inconsistency for horns and wings
In the self-transform menu, the horn and wing types
are sorted before displaying to the player.

The comparator used for sorting, in certain cases,
could reveal itself to violate the general contract.

Since the sort function we are using is stable, the
comparator given to it must also be stable. This is
the general contract.

Stability primarily means that the sort order must be
consistent (for example, if it says [1,2] is in the
correct order, it must then say that [2,1] is in the
incorrect order), and that it must not reorder equal
elements (for example, it must not say that [1,1] is
in the incorrect order).

The existing comparator function would fail this
contract if HornType.NONE (or WingType.NONE) were
ever passed in as the second element.

The rewrite changes from the two-argument comparator
into the single-argument key extractor form, by using
the Comparator class, and first-item enforcement (such
as for HornType.NONE) is performed by mapping the
desired first group to zero, and all other elements
to one.