Camel toes
Updated bulge calculation: The calculation will now account for if the player is a taur. It always bothered me how a two-foot penis could cause a bulge when the taur with said penis is probably longer than two feet

I added descriptions for camel toes in the calculateGenderAppearance method. I additionally added a description to OrificeModifier.PUFFY.