[Bugfix] Command_unknown in sex text
Things to include in a large Pull Request:
What is the purpose of the pull request?
Command_unknown error in starting dialogue from the submissive spitroast(front) with one offspring and one non-offspring.

Give a brief description of what you changed or added.
Seemed that "your" was accidentaly tried as a command and so changed it to just text.

Are any new graphical assets required?
no

Has this change been tested? If so, mention the version number that the test was based on.
4.6

So we have a better idea of who you are, what is your Discord Handle? 
Duskit#6985
First time i did something like this so i'm a bit nervous on whether i did this right.
If you want quick feedback, you can use @Innoxia, or jump on the Lilith's Throne discord and send Innoxia a PM.